Remove updateToken, add attachNewJWT to response object
