+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "derelict",
-  "version": "1.8.0",
+  "version": "1.9.0",
   "description": "Easy Stateless Authentication for Express",
   "main": "dist/index.js",
   "scripts": {
